// scripts/importFromOldDB/libs/API.ts
export enum EmployeeType {
  OTHER = "OTHER",
  NORMAL = "NORMAL",
  MANAGER = "MANAGER",
  SUPER_MANAGER = "SUPER_MANAGER",
}


export type ModelBooleanInput = {
  ne?: boolean | null,
  eq?: boolean | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
};

export type ModelBooleanTypeInput = {
  eq?: BooleanType | null,
  ne?: BooleanType | null,
};

export enum BooleanType {
  TRUE = "TRUE",
  FALSE = "FALSE",
}
// ...
export enum ReportWorkingStatus {
  OK = "OK",
  InTask = "InTask",
  InProblem = "InProblem",
}
// ...
export type CreateEmployeeInput = {
  username: string,
  companyID: string,
  no: string,
  groupID: string,
  superiorUsername: string,
  firstName: string,
  lastName: string,
  grade: string,
  email: string,
  sub?: string | null,
  manager: EmployeeType,
  isCompanyAdmin?: boolean | null,
  isDeleted: BooleanType,
};
// ...
export type CreateGroupInput = {
  id?: string | null,
  companyID: string,
  no: string,
  name: string,
  createdAt?: string | null,
};

export type CreateCategoryInput = {
  id?: string | null,
  companyID: string,
  no: string,
  name: string,
};

export type CreateSheetInput = {
  id?: string | null,
  sub: string,
  year: number,
  companyID: string,
  groupID: string,
  grade: string,
  careerPlan?: string | null,
  careerPlanComment?: string | null,
  reviewComment?: string | null,
  reviewDate?: string | null,
  selfCheckDate?: string | null,
  firstComment?: string | null,
  firstCheckDate?: string | null,
  secondComment?: string | null,
  secondCheckDate?: string | null,
  overAllEvaluation?: number | null,
  statusValue?: number | null,
  interviewPlanDate?: string | null,
  interviewPlanComment?: string | null,
  InterviewMid1Date?: string | null,
  InterviewMid1Comment?: string | null,
  InterviewMid2Date?: string | null,
  InterviewMid2Comment?: string | null,
  InterviewMid3Date?: string | null,
  InterviewMid3Comment?: string | null,
  revieweeUsername: string,
  secondUsername: string,
  sheetGroupName?: string | null,
  referencer?: Array< string > | null,
  reviewee: string,
  topReviewers?: Array< string > | null,
  secondReviewers?: Array< string > | null,
};
// ...
export type CreateSectionInput = {
  id?: string | null,
  sheetID: string,
  no: string,
  sectionCategoryName?: string | null,
  reviewee?: string | null,
  topReviewers?: Array< string > | null,
  secondReviewers?: Array< string > | null,
  referencer?: Array< string > | null,
};

export type CreateObjectiveInput = {
  id?: string | null,
  sectionID: string,
  createdAt?: string | null,
  companyID: string,
  content: string,
  result?: string | null,
  priority?: string | null,
  selfEvaluation?: number | null,
  firstEvaluation?: number | null,
  lastEvaluation?: number | null,
  progress?: number | null,
  expStartDate?: string | null,
  expDoneDate?: string | null,
  reviewee?: string | null,
  topReviewers?: Array< string > | null,
  secondReviewers?: Array< string > | null,
  referencer?: Array< string > | null,
};

export type CreateReportInput = {
  id?: string | null,
  sub: string,
  date: string,
  reviewee?: string | null,
  reviewer?: Array< string | null > | null,
  companyID?: string | null,
  referencer?: Array< string | null > | null,
  revieweeComments?: RevieweeReportCommentInput | null,
  reviewerComments?: ReviewerReportCommentInput | null,
  workStatus?: ReportWorkingStatus | null,
};

export type RevieweeReportCommentInput = {
  work?: string | null,
  other?: string | null,
  status?: string | null,
};

export type ReviewerReportCommentInput = {
  superior?: string | null,
};